Fix sha256 for patch in attr-2.4.48
Adjust sha256 for patches in attr-2.4.48.

When I tried to run fetchurl for a patch in attr-2.4.48 I got an error:

  error: hash mismatch in file downloaded from 'https://git.savannah.nongnu
    wanted: sha256:0gja54fz79a9ma6b4mprnjxq77l5yg2z9xknlwhmkcrfnam02qxp
    got:    sha256:1g8sky52rjg9fpmrddza1af89s9qd4bzsbxqb7mc5bv8gyh10s1k

The error doesn't occur if the attr derivation is taken from cache and
there is no need to build it.

However with https://github.com/NixOS/nix/issues/3369 resolved
`nix-build --check` does a proper check, tries to build from source
and as result the error occurs.